+++ fp2023.cabal
@@ -25,6 +25,8 @@ source-repository head
 
 library
   exposed-modules:
+      DataFrame
+      InMemoryTables
       Lessons.Lesson01
       Lib1
   other-modules:
